Major refactor. All mouse input is forwarded to an InputResponder subclass, and each shape has its own version of said class. Much more effective

+++ cs-355-lab-1/src/cs355/solution/model/shapes/Square.java
@@ -6,7 +6,7 @@ import cs355.solution.util.math.Vector2D;
 
 public class Square extends Rectangle
 {
-	public Square(Color color, Vector2D topLeft, int size)
+	public Square(Color color, Vector2D topLeft, float size)
 	{
 		super(color, topLeft, size, size);
 	}
